Refactor: Adjust stats structure and integrate PlayerStatsDB
- Renamed and restructured properties in `PlayerStats`, `PlayerStatPoint`, and `EntityStats` for consistency.
- Added `PlayerStatsDB` class for database operations related to player stats.
- Updated `ActionbarListener` to reflect the new stat attribute changes.
- Updated Gradle dependencies to latest versions.

+++ src/main/java/kr/hyfata/zero/helper/stats/entity/EntityStats.kt
@@ -1,7 +1,7 @@
 package kr.hyfata.zero.helper.stats.entity
 
 open class EntityStats {
-    var currentHp: Double = 0.0
-    var maxHp: Double = 0.0
+    var currentHp: Int = 0
+    var maxHp: Int = 0
     var level: Int = 0
 }
